test: do not drop database

+++ testing/src/main/kotlin/org/neo4j/connectors/kafka/testing/source/Neo4jSourceExtension.kt
@@ -195,8 +195,10 @@ internal class Neo4jSourceExtension(
       session.close()
       driver.close()
     } else {
-      createDriver().use { dr ->
-        dr.session(SessionConfig.forDatabase("system")).use { it.dropDatabase(neo4jDatabase) }
+      if (!testFailed) {
+        createDriver().use { dr ->
+          dr.session(SessionConfig.forDatabase("system")).use { it.dropDatabase(neo4jDatabase) }
+        }
       }
     }
   }
